Refactor attributes to be re-usable
