+++ tests/Logs/LogTest.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
 namespace Sentry\Tests\Logs;
 
 use PHPUnit\Framework\TestCase;
+use Sentry\Attributes\Attribute;
 use Sentry\Logs\Log;
-use Sentry\Logs\LogAttribute;
 use Sentry\Logs\LogLevel;
 
 /**
- * @phpstan-import-type AttributeValue from LogAttribute
- * @phpstan-import-type AttributeSerialized from LogAttribute
+ * @phpstan-import-type AttributeValue from Attribute
+ * @phpstan-import-type AttributeSerialized from Attribute
  */
 final class LogTest extends TestCase
 {
